refactor: move TooltipProps type definition to a separate file for better organization
