+++ packages/stark-ui/src/components/Tooltip/types.ts
@@ -1,0 +1,8 @@
+export type TooltipProps = {
+  children: React.ReactElement;
+  content: React.ReactNode;
+  placement?: 'top' | 'bottom' | 'left' | 'right';
+  delay?: number;
+  disabled?: boolean;
+  zIndex?: number;
+};
